fix(inputNumber): 마지막에 구분자가 붙어 들어오는 경우가 정상출력되는 오류를 찾음에 따라 오류 수정.
들어오는 값을 InputNumber 로 원시값 포장, stringParser의 검증 로직을 InputNumber로 분리, 도메인 로직수정에 따른 컨트롤러 수정
