+++ README.md
@@ -22,6 +22,14 @@
 
 * [X] 문자열을 입력받는다.
 
+### InputNumbers(들어오는 값 원시값 포장)
+
+* [x] 들어온 값을 원시값 포장한다.
+* [x] 검증 로직을 구현한다.
+    * [x] 빈 값 예외처리를 진행한다.
+    * [x] 마지막에 구분자가 들어있는 경우 예외처리를 진행한다.
+    * [x] 커스텀 구분자의 형식을 지키지 않을 경우 예외처리를 진행한다.
+
 ### 문자열 Parser
 
 * [X] 들어온 값에 따라, 구분자를 추출한다.
@@ -70,7 +78,24 @@
 
 ### 예외처리
 
-* [X] 문자열이 비었을 경우  ```IllegalArgumentException```을 발생시킨다.
-* [X] 커스텀 구분자에 해당하는 문자열의 경우, '\\' 문자를 포함하지 않으면  ```IllegalArgumentException```을 발생시킨다.
-* [X] 숫자 값이 아닌 형태가 있으면 ```IllegalArgumentException```을 발생시킨다.
-* [X] 숫자가 음수이면 ```IllegalArgumentException```을 발생시킨다.
+##### 들어올 수 있는 문자열 : [기본 구분자]
+
+* 빈 값 -> StringParser 에서 예외처리(처리 완료)
+* 1,2;3 기본 구분자 형식 -> 정상 동작(처리 완료)
+* 1,2?3 기본 구분자가 아닌 형식 -> StringCalculator 에서 예외처리(처리 완료)
+* 1,2,a 형식에 다른 문자가 포함된 경우 -> StringCalculator 에서 예외처리(처리 완료)
+* 1,2,3,, 마지막이 기본 구분자가 붙는 경우 -> InputNumber 에서 예외처리(처리 완료)
+
+##### 들어올 수 있는 문자열 : [커스텀 구분자]
+
+* 빈 값 -> StringParser 에서 예외처리(처리 완료)
+* //;\1;2;3 완성되지 않은 형식 -> "\\" 혹은 "/"가 빠질 경우 StringParser 에서 예외처리(처리완료), 다른 문자의 경우 StringCalculator 에서 예외처리(처리완료)
+* //;1n1;2;3 단일 구분자 형식 -> 정상 동작(처리 완료)
+* //;;1;;2;;3 연속된 구분자 형식 -> 정상 동작(처리 완료)
+* //?\n1?2?3 메타 문자가 구분자인 형식 -> 정상동작(처리 완료)
+* //1\n1;2;3 구분자가 숫자인 형식 -> StringCalculator 에서 예외처리(처리 완료)
+* //\n1;2;3 구분자가 없는 경우 -> StringCalculator 에서 예외처리(처리 완료)
+* 1//;\n1;2;3 형식 앞에 숫자가 있는 경우 -> StringCalculator 에서 예외처리 (처리 완료)
+* //;\n1?2?3 구분자가 아닌 문자로 나눠진 경우 -> StringCalculator 에서 예외처리 (처리 완료)
+* //;\n1;2;3;a 마지막에 다른 문자가 붙는 경우 -> StringCalculator 에서 예외처리 (처리 완료)
+* //;\n1;2;3;; 마지막에 구분자가 붙는 경우 -> InputNumber 에서 예외처리(처리 완료)
