+++ docs/수정사항.md
@@ -191,7 +191,7 @@ public String getString(){
 
 * InputView 에서는 입력만 받는 책임을 가져야 하는데, 출력을 하는 문제점을 찾음.
 
-### 해결
+#### 해결
 
 * OutputView
 
@@ -211,3 +211,85 @@ public String getString() {
 
 * 위와 같은 형식으로 메서드를 나눠서 책임에 관한 문제를 해결함.
 * OutputView, InputView 수정에 따라, 컨트롤러도 수정함.
+
+---
+
+### 10/21
+
+#### 문제점 & 분석
+
+* 들어올 수 있는 문자열을 더 분석하여서 예외를 처리하고자 하였음.
+
+##### 들어올 수 있는 문자열 : [기본 구분자]
+
+* 빈 값 -> StringParser 에서 예외처리(처리 완료)
+* 1,2;3 기본 구분자 형식 -> 정상 동작(처리 완료)
+* 1,2?3 기본 구분자가 아닌 형식 -> StringCalculator 에서 예외처리(처리 완료)
+* 1,2,a 형식에 다른 문자가 포함된 경우 -> StringCalculator 에서 예외처리(처리 완료)
+* 1,2,3,, 마지막이 기본 구분자가 붙는 경우 -> 정상출력되어짐. 예외처리 필요(처리 미완료 -> 완료)
+
+##### 들어올 수 있는 문자열 : [커스텀 구분자]
+
+* 빈 값 -> StringParser 에서 예외처리(처리 완료)
+* //;\1;2;3 완성되지 않은 형식 -> "\\" 혹은 "/"가 빠질 경우 StringParser 에서 예외처리(처리완료), 다른 문자의 경우 StringCalculator 에서 예외처리(처리완료)
+* //;1n1;2;3 단일 구분자 형식 -> 정상 동작(처리 완료)
+* //;;1;;2;;3 연속된 구분자 형식 -> 정상 동작(처리 완료)
+* //?\n1?2?3 메타 문자가 구분자인 형식 -> 정상동작(처리 완료)
+* //1\n1;2;3 구분자가 숫자인 형식 -> StringCalculator 에서 예외처리(처리 완료)
+* //\n1;2;3 구분자가 없는 경우 -> StringCalculator 에서 예외처리(처리 완료)
+* 1//;\n1;2;3 형식 앞에 숫자가 있는 경우 -> StringCalculator 에서 예외처리 (처리 완료)
+* //;\n1?2?3 구분자가 아닌 문자로 나눠진 경우 -> StringCalculator 에서 예외처리 (처리 완료)
+* //;\n1;2;3;a 마지막에 다른 문자가 붙는 경우 -> StringCalculator 에서 예외처리 (처리 완료)
+* //;\n1;2;3;; 마지막에 구분자가 붙는 경우 -> 정상출력되어짐. 예외처리 필요(처리 미완료 -> 완료)
+
+**현재 마지막에 구분자가 붙는 경우에 정상출력되어진다는 문제점을 발견하였음.**
+
+#### 해결
+
+* 마지막에 구분자가 포함되는 경우 예외가 발생하도록 하였음.
+* 또한, 들어오는 입력에 대하여 바로 검층 처리를 하고자 하였음. -> 왜냐하면 현재 StringParser 클래스에 검증 로직이 있는데, 검증 로직을 따로 구분하여 관심사를 분리하고자 하였음.
+* 검증 로직 분리 해결 방안 -> 원시값 포장: InputNumbers 로 String inputString 을 원시값 포장하여 사용.
+
+**InputNumber**
+
+~~~
+package calculator.domain;
+
+public class InputNumber {
+
+    private static final String EXCEPTION_EMPTY_STRING = "빈 값을 입력하셨습니다.";
+    private static final String EXCEPTION_CUSTOM_STRING = "올바른 형식으로 입력하세요.";
+    private static final String CUSTOM_NUMBER_PART_DELIMITER = "\\";
+
+    private final String inputStrings;
+
+    public InputNumber(String inputStrings) {
+        validateEmptyString(inputStrings);
+        validateLastIndex(inputStrings);
+        validateCustomStrings(inputStrings);
+        this.inputStrings = inputStrings;
+    }
+
+    public String getInputStrings() {
+        return inputStrings;
+    }
+
+    private void validateEmptyString(String inputStrings) {
+        if (inputStrings.isEmpty()) {
+            throw new IllegalArgumentException(EXCEPTION_EMPTY_STRING);
+        }
+    }
+
+    private void validateCustomStrings(String inputStrings) {
+        if (!inputStrings.contains(CUSTOM_NUMBER_PART_DELIMITER)) {
+            throw new IllegalArgumentException(EXCEPTION_CUSTOM_STRING);
+        }
+    }
+
+    private void validateLastIndex(String inputStrings) {
+        if (!Character.isDigit(inputStrings.charAt(inputStrings.length() - 1))) {
+            throw new IllegalArgumentException(EXCEPTION_CUSTOM_STRING);
+        }
+    }
+}
+~~~
